requirements.txt
fastapi
uvicorn
jinja2
sd-parsers
alive-progress
pillow
pytest
pytest-cov
